# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Book Store",
        "description": "API para gestionar libros, autores, editoriales y categorias",
        "contact": {
            "email": "angelj.vazquez1@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/sales": {
            "post": {
                "tags": [
                    "Sales"
                ],
                "summary": "Crear una nueva venta con detalles",
                "operationId": "0e806786857f022e42a4097cb3e644ae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_client",
                                    "saleDetails"
                                ],
                                "properties": {
                                    "id_client": {
                                        "description": "ID del cliente que realiza la venta",
                                        "type": "integer",
                                        "example": 123
                                    },
                                    "saleDetails": {
                                        "description": "Lista de detalles de la venta",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "book_id",
                                                "quantity",
                                                "unit_price"
                                            ],
                                            "properties": {
                                                "book_id": {
                                                    "description": "ID del libro vendido",
                                                    "type": "integer",
                                                    "example": 456
                                                },
                                                "quantity": {
                                                    "description": "Cantidad vendida del libro",
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "unit_price": {
                                                    "description": "Precio unitario del libro",
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 19.99
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Venta creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sale created successfully"
                                        },
                                        "sale": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id_client": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-02T12:34:56Z"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 39.98
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear la venta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 500
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sale creation failed"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Error message describing the failure"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sales"
                ],
                "summary": "Eliminar una venta por su ID",
                "operationId": "e01c1d64ae79bdef0737ca0f536fe5d8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "query",
                        "description": "ID de la venta a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venta eliminada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sale deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Venta no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sale not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sales/{id_sale}": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Obtener una venta por su ID con detalles",
                "operationId": "ca5f8f4083141997afd52b7ccf697130",
                "parameters": [
                    {
                        "name": "id_sale",
                        "in": "path",
                        "description": "ID de la venta a obtener",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Venta encontrada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sale found successfully"
                                        },
                                        "sale": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "id_client": {
                                                    "type": "integer",
                                                    "example": 123
                                                },
                                                "date": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2025-06-02T12:34:56Z"
                                                },
                                                "total": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 39.98
                                                },
                                                "sale_detail": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 10
                                                            },
                                                            "id_sale": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "id_book": {
                                                                "type": "integer",
                                                                "example": 456
                                                            },
                                                            "quantity": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "unit_price": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 19.99
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Venta no encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sale not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sales/by-user": {
            "get": {
                "tags": [
                    "Sales"
                ],
                "summary": "Obtener ventas por ID de usuario",
                "operationId": "cab26c65e03ed5546de534d473cba439",
                "parameters": [
                    {
                        "name": "id_user",
                        "in": "query",
                        "description": "ID del usuario (cliente) para obtener sus ventas",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 123
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ventas encontradas exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Sale found successfully"
                                        },
                                        "sales": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Sale"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ventas no encontradas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Sale not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "SaleDetail": {
                "title": "SaleDetail",
                "description": "Detalle de una venta",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "id_sale": {
                        "type": "integer",
                        "example": 10
                    },
                    "id_book": {
                        "type": "integer",
                        "example": 5
                    },
                    "quantity": {
                        "type": "integer",
                        "example": 2
                    },
                    "unit_price": {
                        "type": "number",
                        "format": "float",
                        "example": 15.5
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "Sale": {
                "title": "Sale",
                "description": "Venta realizada por un cliente",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 10
                    },
                    "id_client": {
                        "type": "integer",
                        "example": 123
                    },
                    "date": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2025-06-02T15:30:00Z"
                    },
                    "total": {
                        "type": "number",
                        "format": "float",
                        "example": 50
                    },
                    "sale_details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SaleDetail"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}